push_service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
django_celery_results==2.6.0
djangorestframework==3.16.1
firebase-admin==6.5.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
google-auth==2.43.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
multidict==6.7.0
packaging==25.0
pika==1.3.2
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.1
pyfcm==2.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
